Adicionadas as classes de divisão e raiz quadrão ao padrão interpreter, bem como atualizada a interface IExpressao para o método Avalia retornar double ao invés de int

+++ DesignPatterns/Interpreter/Calculadora/Interfaces/IExpressao.cs
@@ -8,6 +8,6 @@ namespace Calculadora.Interfaces
 {
     internal interface IExpressao
     {
-        public int Avalia();
+        public double Avalia();
     }
 }
